requirements.txt

# Core
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Web Scraping
crawl4ai>=0.4.0

# Data
pydantic>=2.7.0
pandas>=2.0.0

# Config
python-dotenv>=1.0.0

# Search
tavily>=0.2.0

# UI
gradio>=4.0.0

# HTTP
httpx>=0.27.0
